Added google tests

+++ symmetrykit/skparser.cpp
@@ -28,9 +28,9 @@ std::vector<std::vector<std::shared_ptr<SKStructure>>> SKParser::movies()
     return _movies;
 }
 
-std::vector<std::tuple<double3, int, double> > SKParser::firstTestFrame()
+std::vector<std::tuple<double3, size_t, double> > SKParser::firstTestFrame()
 {
-    std::vector<std::tuple<double3, int, double> > atoms{};
+    std::vector<std::tuple<double3, size_t, double> > atoms{};
 
     for (const std::vector<std::shared_ptr<SKStructure>>& movie : _movies)
     {
@@ -38,7 +38,7 @@ std::vector<std::tuple<double3, int, double> > SKParser::firstTestFrame()
         {
             for (const std::shared_ptr<SKAsymmetricAtom>& atom : structure->atoms)
             {
-                std::tuple<double3, int, double> atomTuple = std::make_tuple<double3, int, double>(atom->position(), static_cast<int>(atom->elementIdentifier()), 1.0);
+                std::tuple<double3, size_t, double> atomTuple = std::make_tuple<double3, size_t, double>(atom->position(), static_cast<size_t>(atom->elementIdentifier()), 1.0);
                 atoms.push_back(atomTuple);
             }
         }
